+++ components/FlexibleHeader/examples/FlexibleHeaderConfiguratorExample.m
@@ -95,6 +95,10 @@
     case FlexibleHeaderConfiguratorFieldMaximumHeight:
       headerView.maximumHeight = [self heightDenormalized:[value floatValue]];
       break;
+
+    case FlexibleHeaderConfiguratorFieldMinMaxHeightIncludeSafeArea:
+      headerView.minMaxHeightIncludesSafeArea = [value boolValue];
+      break;
   }
 }
 
@@ -209,6 +213,8 @@ static const CGFloat kHeightScalar = 300;
 
     case FlexibleHeaderConfiguratorFieldMaximumHeight:
       return @([self normalizedHeight:self.fhvc.headerView.maximumHeight]);
+    case FlexibleHeaderConfiguratorFieldMinMaxHeightIncludeSafeArea:
+      return @(self.fhvc.headerView.minMaxHeightIncludesSafeArea);
   }
 }